predict.py major error fix

+++ modules/dataloaders/utils.py
@@ -41,7 +41,7 @@ def decode_segmap(label_mask, dataset, plot=False):
         r[label_mask == ll] = label_colours[ll, 0]
         g[label_mask == ll] = label_colours[ll, 1]
         b[label_mask == ll] = label_colours[ll, 2]
-    rgb = np.zeros((label_mask.shape[0], label_mask.shape[1], 3))
+    rgb = np.zeros((label_mask.shape[0], label_mask.shape[1], 3), dtype=np.float32)
     rgb[:, :, 0] = r / 255.0
     rgb[:, :, 1] = g / 255.0
     rgb[:, :, 2] = b / 255.0
